:bug: detect storage to sync state close #32
